v1.0.3 Split this functionality into two separate UI Scripts to work around ServiceNow's weird and broken handling of JS Includes when two UI Scripts have the same name. Portal behavior should now be much more stable. For real this time.
v1.0.2
Fixed an issue with scope binding specifically in Service Portal.
Added a JS Include to the default "Stock" Service Portal theme to include the minified version of EfficientGlideRecord. If you use a different theme for your Service Portal, please be sure to add a JS Include for the portal version of the EfficientGlideRecord UI Script to your theme.

v1.0.1
Added .getRowCount() method to get the number of rows returned from the server (limited by .setLimit() if specified).
Improved documentation and examples (in this article as well as in the JSDoc in the code).
Updated get(recordSysID) so you no longer have to call next() in the callback function (it should've been this way from the start; my bad).

v1.0.0
Initial public release